split demo.c test suite
